Add LRU cache
