+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/krakendio/httpcache
 
 go 1.19
+
+require github.com/hashicorp/golang-lru v1.0.2
