+++ test/test_test.go
@@ -10,3 +10,7 @@ import (
 func TestMemoryCache(t *testing.T) {
 	test.Cache(t, httpcache.NewMemoryCache())
 }
+
+func TestLruCache(t *testing.T) {
+	test.Cache(t, httpcache.NewLruCache(10))
+}
